refactor(tests): address review feedback
This commit refactors the test suite to address code review feedback.

- **DRY Golden File Helpers:** The duplicated golden file helper functions (`get_file_hash`, `check_golden_file_hash`) have been moved to a single shared file, `tests/helpers.py`, to adhere to the DRY principle. `test_charts.py` and `test_report.py` now import the helpers from this shared location.

- **Safer Environment Variable Handling:** The report test (`tests/test_report.py`) now uses the `monkeypatch` fixture to safely set and tear down the `SOURCE_DATE_EPOCH` environment variable, preventing potential test leakage.

- **Test Package:** An `__init__.py` file has been added to the `tests/` directory to ensure it is treated as a package, allowing for clean imports of the shared helper module.
